+++ src/main/java/roomescape/repository/reservationTime/JdbcReservationTimeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.repository.reservationTime;
 
 import java.util.Collections;
 import java.util.List;
@@ -9,25 +9,29 @@ import roomescape.domain.ReservationTime.ReservationTime;
 import roomescape.domain.ReservationTime.ReservationTimeCommand;
 
 @Repository
-public class ReservationTimeRepository {
+public class JdbcReservationTimeRepository implements ReservationTimeRepository {
     private final ReservationTimeDao reservationTimeDao;
 
-    public ReservationTimeRepository(ReservationTimeDao reservationTimeDao) {
+    public JdbcReservationTimeRepository(ReservationTimeDao reservationTimeDao) {
         this.reservationTimeDao = reservationTimeDao;
     }
 
+    @Override
     public ReservationTime addReservationTime(ReservationTimeCommand reservationTimeCommand) {
         return new ReservationTime(reservationTimeDao.insertReservationTime(reservationTimeCommand), reservationTimeCommand.startAt());
     }
 
+    @Override
     public Optional<ReservationTime> getReservationTime(long id) {
         return reservationTimeDao.getReservationTime(id);
     }
 
+    @Override
     public List<ReservationTime> getAllReservationTime() {
         return Collections.unmodifiableList(reservationTimeDao.getAllReservationTime());
     }
 
+    @Override
     public void deleteReservationTime(long id) {
         reservationTimeDao.deleteReservationTime(id);
     }
