+++ src/main/java/roomescape/repository/reservationTime/ReservationTimeRepository.java
@@ -1,0 +1,13 @@
+package roomescape.repository.reservationTime;
+
+import java.util.List;
+import java.util.Optional;
+import roomescape.domain.ReservationTime.ReservationTime;
+import roomescape.domain.ReservationTime.ReservationTimeCommand;
+
+public interface ReservationTimeRepository {
+    ReservationTime addReservationTime(ReservationTimeCommand reservationTimeCommand);
+    Optional<ReservationTime> getReservationTime(long id);
+    List<ReservationTime> getAllReservationTime();
+    void deleteReservationTime(long id);
+}
